Fixing preview endpoint + fixes in metadata
