+++ src/app/api/preview/route.ts
@@ -8,6 +8,7 @@ import { sanityFetch } from "@/lib/backend/sanity/client";
 
 // Queries
 import { VALIDATE_DOCUMENT_BY_SLUG_QUERY } from "@/lib/backend/sanity/queries";
+import { buildPath } from "@/lib/backend/sanity/utils";
 
 // Constants
 const SECRET = process.env.NEXT_SANITY_PREVIEW_SECRET;
@@ -22,12 +23,13 @@ export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
   const secret = searchParams.get("secret");
   const slug = searchParams.get("slug");
+  const type = searchParams.get("type");
 
   if (secret !== SECRET) {
     return new Response("Invalid token", { status: 401 });
   }
 
-  if (!slug) {
+  if (!slug || !type) {
     return new Response("Invalid payload", { status: 401 });
   }
 
@@ -48,7 +50,8 @@ export async function GET(request: Request) {
   draftMode().enable();
 
   // Redirect to the current preview page
-  redirect(`/${_slug}`);
+  const path = buildPath(slug, type);
+  redirect(path);
 }
 
 export const OPTIONS = async () => {
